+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.10
 
+* Support multiple tunnels on same host
 * Dev only: more explicit typing for IP addresses throughout
 
 ## 0.2.9